Refactor artifact package structure
