+++ optuna_dashboard/_serializer.py
@@ -12,8 +12,8 @@ from optuna.study import StudySummary
 from optuna.trial import FrozenTrial
 
 from . import _note as note
-from . import artifact
 from ._named_objectives import get_objective_names
+from .artifact._backend import list_trial_artifacts
 
 
 if TYPE_CHECKING:
@@ -180,7 +180,7 @@ def serialize_frozen_trial(
             {k: trial_system_attrs[k] for k in trial_system_attrs if not k.startswith("dashboard")}
         ),
         "note": note.get_note_from_system_attrs(study_system_attrs, trial._trial_id),
-        "artifacts": artifact._list_trial_artifacts(study_system_attrs, trial._trial_id),
+        "artifacts": list_trial_artifacts(study_system_attrs, trial._trial_id),
     }
 
     serialized_intermediate_values: list[IntermediateValue] = []
